requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit
joblib
osmnx
geopy
requestsstreamlit run frontend.py
